requirements.txt
Flask
websockets
paho-mqtt